requirements.txt
apex==0.9.10dev
einops==0.3.0
ftfy==5.8
ipython==8.11.0
lmdb==1.4.0
matplotlib==3.4.3
mnist==0.2.2
nltk==3.7
numpy==1.21.5
opencv_python_headless==4.7.0.68
Pillow==9.4.0
regex==2021.11.2
scikit_learn==1.2.2
scipy==1.7.1
setuptools==52.0.0.post20210125
stl10==0.0.4
svhn==0.0.1
timm==0.6.12
torchvision==0.8.2
tqdm==4.64.1
transformers
